requirements.txt
flask
flask-restful
Flask-PyMongo
pymongo
flask_httpauth
pyjwt